Rename setup -> modules
