+++ packages/backend/src/Server.ts
@@ -19,6 +19,8 @@ import { BlockNumberUpdater } from './core/BlockNumberUpdater'
 import { Clock } from './core/Clock'
 import { PriceUpdater } from './core/PriceUpdater'
 import { ReportUpdater } from './core/reports/ReportUpdater'
+import { getActivityModule } from './modules/activity/ActivityModule'
+import { getActivityV2Module } from './modules/activityV2/ActivityV2Module'
 import { CoingeckoQueryService } from './peripherals/coingecko/CoingeckoQueryService'
 import { AggregateReportRepository } from './peripherals/database/AggregateReportRepository'
 import { BalanceRepository } from './peripherals/database/BalanceRepository'
@@ -31,8 +33,6 @@ import { Database } from './peripherals/database/shared/Database'
 import { EthereumClient } from './peripherals/ethereum/EthereumClient'
 import { MulticallClient } from './peripherals/ethereum/MulticallClient'
 import { EtherscanClient } from './peripherals/etherscan'
-import { getActivityModule } from './setup/ActivityModule'
-import { getActivityV2Module } from './setup/ActivityV2Module'
 import { handleServerError, reportError } from './tools/ErrorReporter'
 
 export class Server {
