+++ packages/backend/src/modules/activity/ActivityModule.ts
@@ -1,21 +1,21 @@
 import { HttpClient, Logger } from '@l2beat/common'
 import { ProjectId } from '@l2beat/types'
 
-import { ActivityController } from '../api/controllers/activity/ActivityController'
-import { createActivityRouter } from '../api/routers/ActivityRouter'
-import { Config } from '../config'
-import { Clock } from '../core/Clock'
-import { LoopringTransactionUpdater } from '../core/transaction-count/LoopringTransactionUpdater'
-import { MaterializedViewRefresher } from '../core/transaction-count/MaterializedViewRefresher'
-import { ZksyncTransactionUpdater } from '../core/transaction-count/ZksyncTransactionUpdater'
-import { BlockTipRepository } from '../peripherals/database/BlockTipRepository'
-import { BlockTransactionCountRepository } from '../peripherals/database/BlockTransactionCountRepository'
-import { Database } from '../peripherals/database/shared/Database'
-import { StarkexTransactionCountRepository } from '../peripherals/database/StarkexTransactionCountRepository'
-import { ZksyncTransactionRepository } from '../peripherals/database/ZksyncTransactionRepository'
-import { LoopringClient } from '../peripherals/loopring'
-import { StarkexClient } from '../peripherals/starkex'
-import { ZksyncClient } from '../peripherals/zksync'
+import { ActivityController } from '../../api/controllers/activity/ActivityController'
+import { createActivityRouter } from '../../api/routers/ActivityRouter'
+import { Config } from '../../config'
+import { Clock } from '../../core/Clock'
+import { LoopringTransactionUpdater } from '../../core/transaction-count/LoopringTransactionUpdater'
+import { MaterializedViewRefresher } from '../../core/transaction-count/MaterializedViewRefresher'
+import { ZksyncTransactionUpdater } from '../../core/transaction-count/ZksyncTransactionUpdater'
+import { BlockTipRepository } from '../../peripherals/database/BlockTipRepository'
+import { BlockTransactionCountRepository } from '../../peripherals/database/BlockTransactionCountRepository'
+import { Database } from '../../peripherals/database/shared/Database'
+import { StarkexTransactionCountRepository } from '../../peripherals/database/StarkexTransactionCountRepository'
+import { ZksyncTransactionRepository } from '../../peripherals/database/ZksyncTransactionRepository'
+import { LoopringClient } from '../../peripherals/loopring'
+import { StarkexClient } from '../../peripherals/starkex'
+import { ZksyncClient } from '../../peripherals/zksync'
 import { createAztecTransactionUpdaters } from './createAztecTransactionUpdaters'
 import {
   createEthereumTransactionUpdater,
