+++ packages/backend/src/modules/activity/createAztecTransactionUpdaters.ts
@@ -1,10 +1,10 @@
 import { HttpClient, Logger } from '@l2beat/common'
 
-import { Config } from '../config/Config'
-import { Clock } from '../core/Clock'
-import { AztecTransactionUpdater } from '../core/transaction-count/AztecTransactionUpdater'
-import { AztecClient } from '../peripherals/aztec'
-import { BlockTransactionCountRepository } from '../peripherals/database/BlockTransactionCountRepository'
+import { Config } from '../../config/Config'
+import { Clock } from '../../core/Clock'
+import { AztecTransactionUpdater } from '../../core/transaction-count/AztecTransactionUpdater'
+import { AztecClient } from '../../peripherals/aztec'
+import { BlockTransactionCountRepository } from '../../peripherals/database/BlockTransactionCountRepository'
 
 export function createAztecTransactionUpdaters(
   httpClient: HttpClient,
