+++ packages/backend/src/modules/activity/createRpcTransactionUpdaters.ts
@@ -2,13 +2,13 @@ import { assert, HttpClient, Logger } from '@l2beat/common'
 import { ProjectId } from '@l2beat/types'
 import { providers } from 'ethers'
 
-import { Config } from '../config'
-import { TransactionCountSyncConfig } from '../config/Config'
-import { Clock } from '../core/Clock'
-import { RpcTransactionUpdater } from '../core/transaction-count/RpcTransactionUpdater'
-import { BlockTransactionCountRepository } from '../peripherals/database/BlockTransactionCountRepository'
-import { EthereumClient } from '../peripherals/ethereum/EthereumClient'
-import { StarkNetClient } from '../peripherals/starknet/StarkNetClient'
+import { Config } from '../../config'
+import { TransactionCountSyncConfig } from '../../config/Config'
+import { Clock } from '../../core/Clock'
+import { RpcTransactionUpdater } from '../../core/transaction-count/RpcTransactionUpdater'
+import { BlockTransactionCountRepository } from '../../peripherals/database/BlockTransactionCountRepository'
+import { EthereumClient } from '../../peripherals/ethereum/EthereumClient'
+import { StarkNetClient } from '../../peripherals/starknet/StarkNetClient'
 
 export function createLayer2RpcTransactionUpdaters(
   config: Config,
