+++ packages/backend/src/modules/activityV2/ActivityV2Module.ts
@@ -1,14 +1,14 @@
 import { HttpClient, Logger } from '@l2beat/common'
 import { ProjectId } from '@l2beat/types'
 
-import { ActivityV2Controller } from '../api/controllers/ActivityV2Controller'
-import { createActivityV2Router } from '../api/routers/ActivityV2Router'
-import { Config } from '../config'
-import { DailyTransactionCountService } from '../core/activity/DailyTransactionCountService'
-import { Clock } from '../core/Clock'
-import { SequenceProcessor } from '../core/SequenceProcessor'
-import { DailyTransactionCountViewRepository } from '../peripherals/database/activity-v2/DailyTransactionCountRepository'
-import { Database } from '../peripherals/database/shared/Database'
+import { ActivityV2Controller } from '../../api/controllers/ActivityV2Controller'
+import { createActivityV2Router } from '../../api/routers/ActivityV2Router'
+import { Config } from '../../config'
+import { DailyTransactionCountService } from '../../core/activity/DailyTransactionCountService'
+import { Clock } from '../../core/Clock'
+import { SequenceProcessor } from '../../core/SequenceProcessor'
+import { DailyTransactionCountViewRepository } from '../../peripherals/database/activity-v2/DailyTransactionCountRepository'
+import { Database } from '../../peripherals/database/shared/Database'
 import { createSequenceProcessors } from './createSequenceProcessors'
 
 export function getActivityV2Module(
