+++ packages/backend/src/modules/activityV2/createSequenceProcessors.ts
@@ -3,21 +3,21 @@ import { assert } from '@l2beat/common/src/tools/assert'
 import { Layer2TransactionApi } from '@l2beat/config'
 import { ProjectId } from '@l2beat/types'
 
-import { Config } from '../config'
-import { createAztecProcessor } from '../core/activity/processors/AztecProcessor'
-import { createLoopringProcessor } from '../core/activity/processors/LoopringProcessor'
-import { createRpcProcessor } from '../core/activity/processors/RpcProcessor'
-import { createStarkexProcessor } from '../core/activity/processors/StarkexProcessor'
-import { createStarknetProcessor } from '../core/activity/processors/StarknetProcessor'
-import { createZksyncProcessor } from '../core/activity/processors/ZksyncProcessor'
-import { Clock } from '../core/Clock'
-import { SequenceProcessor } from '../core/SequenceProcessor'
-import { Project } from '../model'
-import { BlockTransactionCountRepository } from '../peripherals/database/activity-v2/BlockTransactionCountRepository'
-import { StarkexTransactionCountRepository } from '../peripherals/database/activity-v2/StarkexCountRepository'
-import { SequenceProcessorRepository } from '../peripherals/database/SequenceProcessorRepository'
-import { Database } from '../peripherals/database/shared/Database'
-import { StarkexClient } from '../peripherals/starkex'
+import { Config } from '../../config'
+import { createAztecProcessor } from '../../core/activity/processors/AztecProcessor'
+import { createLoopringProcessor } from '../../core/activity/processors/LoopringProcessor'
+import { createRpcProcessor } from '../../core/activity/processors/RpcProcessor'
+import { createStarkexProcessor } from '../../core/activity/processors/StarkexProcessor'
+import { createStarknetProcessor } from '../../core/activity/processors/StarknetProcessor'
+import { createZksyncProcessor } from '../../core/activity/processors/ZksyncProcessor'
+import { Clock } from '../../core/Clock'
+import { SequenceProcessor } from '../../core/SequenceProcessor'
+import { Project } from '../../model'
+import { BlockTransactionCountRepository } from '../../peripherals/database/activity-v2/BlockTransactionCountRepository'
+import { StarkexTransactionCountRepository } from '../../peripherals/database/activity-v2/StarkexCountRepository'
+import { SequenceProcessorRepository } from '../../peripherals/database/SequenceProcessorRepository'
+import { Database } from '../../peripherals/database/shared/Database'
+import { StarkexClient } from '../../peripherals/starkex'
 
 export function createSequenceProcessors(
   config: Config,
